--- src/app/(pages)/auth/sign-up/page.tsx
@@ -1,7 +1,0 @@
-import * as React from 'react';
-
-import { SignUpForm } from '@/features/signUp/components';
-
-export default function SignUpPage() {
-  return <SignUpForm />;
-}
