--- src/app/(pages)/projects/[projectId]/layout.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import { PyodideProvider } from '@/lib/pyodide';
-
-export default function ProjectPageLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return <PyodideProvider>{children}</PyodideProvider>;
-}
